requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
sqlalchemy>=2.0.28
asyncpg>=0.29.0
python-telegram-bot>=21.0.1
gspread>=6.0.2
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
apscheduler>=3.10.4
aiosqlite>=0.20.0
httpx>=0.27.0
pyngrok>=7.1.0
